Organized, documented and cleaned up code
